requirements-dev.txt
mypy==0.760
flake8==3.7.*
numpy
ipython
sphinx-autodoc-typehints
sphinx_rtd_theme
autopep8
flake8-docstrings
flake8-bugbear
flake8-commas
flake8-colors
pep8-naming
pydocstyle
pylint
yapf
pudb
black
pytest-pudb
pytest-cov
versioneer
